app-python/PruebaPython/requirements.txt
hyperliquid-python-sdk
eth-account
requests 